[hotfix][test] Deduplicate code of ValidatingCheckpointHandler
